# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ryden API
  version: 0.24.0
  description: Authenticated coordination through either shared planning or a fixed plan, with a private friend address book that never grants meeting access by itself.
servers:
  - url: /api/v1
paths:
...
  /friend-requests:
    post:
      operationId: sendFriendRequest
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
      responses:
        "201": { description: Request created, content: { application/json: { schema: { $ref: "#/components/schemas/ChangedMutation" } } } }
        "200": { description: An equivalent relationship already exists, content: { application/json: { schema: { $ref: "#/components/schemas/ChangedMutation" } } } }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
...
  /friends/{friendID}:
    delete:
      operationId: removeFriend
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: friendID
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Friendship removed. }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: Structured application error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    ChangedMutation:
      type: object
      required: [changed]
      properties:
        changed: { type: boolean }
...
    Problem:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }